Moving available datastores from DatastoreCatalog to DatabaseDriverCatalog.
