+++ desktop/ui/src/main/java/org/datacleaner/panels/SelectDatastoreContainerPanel.java
@@ -27,6 +27,7 @@ import javax.swing.JScrollPane;
 import javax.swing.SwingUtilities;
 
 import org.datacleaner.connection.Datastore;
+import org.datacleaner.database.DatabaseDriverCatalog;
 import org.datacleaner.guice.DCModule;
 import org.datacleaner.user.DatastoreChangeListener;
 import org.datacleaner.user.DatastoreSelectedListener;
@@ -46,12 +47,12 @@ public class SelectDatastoreContainerPanel extends DCSplashPanel implements Data
     private final SelectDatastorePanel _selectDatastorePanel;
     private final MutableDatastoreCatalog _datastoreCatalog;
 
-    public SelectDatastoreContainerPanel(AnalysisJobBuilderWindow window, DCModule dcModule, MutableDatastoreCatalog datastoreCatalog,
+    public SelectDatastoreContainerPanel(AnalysisJobBuilderWindow window, DCModule dcModule, DatabaseDriverCatalog databaseDriverCatalog, MutableDatastoreCatalog datastoreCatalog,
             UserPreferences userPreferences) {
         super(window);
         _datastoreCatalog = datastoreCatalog;
         _selectDatastorePanel = new SelectDatastorePanel(dcModule, datastoreCatalog,
-                userPreferences, this, true);
+                databaseDriverCatalog, userPreferences, this, true);
 
         setLayout(new BorderLayout());
         final JScrollPane scroll = wrapContent(_selectDatastorePanel);
