+++ desktop/ui/src/main/java/org/datacleaner/panels/SelectDatastorePanel.java
@@ -23,6 +23,7 @@ import javax.swing.Box;
 import javax.swing.JComponent;
 
 import org.datacleaner.connection.DatastoreCatalog;
+import org.datacleaner.database.DatabaseDriverCatalog;
 import org.datacleaner.guice.DCModule;
 import org.datacleaner.user.DatastoreSelectedListener;
 import org.datacleaner.user.UserPreferences;
@@ -34,7 +35,7 @@ public class SelectDatastorePanel extends DCPanel {
 
     private final ExistingDatastorePanel _existingDatastoresPanel;
 
-    public SelectDatastorePanel(DCModule dcModule, DatastoreCatalog datastoreCatalog, UserPreferences userPreferences,
+    public SelectDatastorePanel(DCModule dcModule, DatastoreCatalog datastoreCatalog, DatabaseDriverCatalog databaseDriverCatalog, UserPreferences userPreferences,
             DatastoreSelectedListener datastoreSelectListener, boolean showExistingDatastoresAsLongList) {
         super();
         setLayout(new VerticalLayout());
@@ -48,7 +49,7 @@ public class SelectDatastorePanel extends DCPanel {
             add(newDatastoreLabel);
         }
 
-        add(new AddDatastorePanel(datastoreCatalog, dcModule, datastoreSelectListener, userPreferences,
+        add(new AddDatastorePanel(datastoreCatalog, databaseDriverCatalog, dcModule, datastoreSelectListener, userPreferences,
                 !showExistingDatastoresAsLongList));
 
         if (showExistingDatastoresAsLongList) {
